Additional ExtendedAE-Plus compatibility
Don't use MixinCraftingCPUCluster because ExtendedAE-Plus does the same thing and both mods used redirects so we got a conflict.
Disabled the mixin here and not in ExtendedAE since that one sets a new thread limit (even if it's int.max) and doesn't just fake a calculation result like this mod's mixin does.